requirements.txt
python-rapidjson
tqdm
requests
git
pygithub
joblib
bzip2
